+++ src/mefit/Common/Constants/ApplicationUrls.cs
@@ -1,0 +1,22 @@
+﻿// Mac EFI Toolkit
+// https://github.com/MuertoGB/MacEfiToolkit
+
+// ApplicationUrls.cs
+// Released under the GNU GLP v3.0
+
+namespace Mac_EFI_Toolkit.Common.Constants
+{
+    internal static class ApplicationUrls
+    {
+        internal const string Changelog = "https://github.com/MuertoGB/MacEfiToolkit/blob/main/CHANGELOG.md";
+        internal const string PaypalDonate = "https://www.paypal.com/donate/?hosted_button_id=Z88F3UEZB47SQ";
+        internal const string Email = "mailto:muertogb@proton.me";
+        internal const string FlexBv5 = "https://pldaniels.com/flexbv5/";
+        internal const string GithubHomepage = "https://github.com/MuertoGB/MacEfiToolkit";
+        internal const string GithubManual = "https://github.com/MuertoGB/MacEfiToolkit/blob/main/MANUAL.md#application-manual";
+        internal const string GithubIssues = "https://github.com/MuertoGB/MacEfiToolkit/issues";
+        internal const string GithubLatestVersion = "https://github.com/MuertoGB/MacEfiToolkit/releases/latest";
+        internal const string VersionManifest = "https://raw.githubusercontent.com/MuertoGB/MacEfiToolkit/main/stream/manifests/version.xml";
+
+    }
+}
