+++ src/mefit/Common/Constants/ApplicationVersions.cs
@@ -1,0 +1,15 @@
+﻿// Mac EFI Toolkit
+// https://github.com/MuertoGB/MacEfiToolkit
+
+// ApplicationVersions.cs
+// Released under the GNU GLP v3.0
+
+namespace Mac_EFI_Toolkit.Common.Constants
+{
+    internal static class ApplicationVersions
+    {
+        internal const string LZMA_SDK_VERSION = "24.08";
+        internal const string CURRENT_BUILD = "250306.1145";
+        internal const string CURRENT_CHANNEL = "Stable";
+    }
+}
