+++ src/mefit/Forms/frmRomInfo.cs
@@ -5,6 +5,7 @@
 // frmRomInfo.cs
 // Released under the GNU GLP v3.0
 
+using Mac_EFI_Toolkit.Common.Constants;
 using Mac_EFI_Toolkit.Firmware.EFIROM;
 using Mac_EFI_Toolkit.Tools;
 using Mac_EFI_Toolkit.UI;
