+++ src/mefit/Forms/frmSettings.cs
@@ -5,6 +5,7 @@
 // frmSettings.cs
 // Released under the GNU GLP v3.0
 
+using Mac_EFI_Toolkit.Common.Constants;
 using Mac_EFI_Toolkit.UI;
 using System;
 using System.IO;
