+++ src/mefit/Logger.cs
@@ -4,6 +4,7 @@
 // Logger.cs - Handles logging of data and text
 // Released under the GNU GLP v3.0
 
+using Mac_EFI_Toolkit.Common.Constants;
 using Mac_EFI_Toolkit.UI;
 using System;
 using System.Diagnostics;
