+++ src/mefit/Program.cs
@@ -6,10 +6,10 @@
 // MET uses embedded font resource "Segoe MDL2 Assets" which is copyright Microsoft Corp.
 
 using Mac_EFI_Toolkit.Common;
+using Mac_EFI_Toolkit.Common.Constants;
 using Mac_EFI_Toolkit.Forms;
 using Mac_EFI_Toolkit.Tools;
 using Mac_EFI_Toolkit.UI;
-using Mac_EFI_Toolkit.WIN32;
 using System;
 using System.ComponentModel;
 using System.Diagnostics;
@@ -22,46 +22,7 @@ using System.Xml;
 
 namespace Mac_EFI_Toolkit
 {
-    #region Struct
-    internal readonly struct ApplicationPaths
-    {
-        internal static readonly string WorkingDirectory = AppDomain.CurrentDomain.BaseDirectory;
-        internal static readonly string FriendlyName = AppDomain.CurrentDomain.FriendlyName;
-        internal static readonly string BackupsDirectory = Path.Combine(WorkingDirectory, "backups");
-        internal static readonly string BuildsDirectory = Path.Combine(WorkingDirectory, "builds");
-        internal static readonly string FsysDirectory = Path.Combine(WorkingDirectory, "fsys_stores");
-        internal static readonly string IntelMeDirectory = Path.Combine(WorkingDirectory, "me_regions");
-        internal static readonly string NvramDirectory = Path.Combine(WorkingDirectory, "nvram_stores");
-        internal static readonly string ScfgDirectory = Path.Combine(WorkingDirectory, "scfg_stores");
-        internal static readonly string LzmaDirectory = Path.Combine(WorkingDirectory, "lzma_archives");
-        internal static readonly string SettingsFile = Path.Combine(WorkingDirectory, "Settings.ini");
-        internal static readonly string ApplicationLog = Path.Combine(WorkingDirectory, "application.log");
-        internal static readonly string DatabaseLog = Path.Combine(WorkingDirectory, "database.log");
-    }
-
-    internal readonly struct ApplicationUrls
-    {
-        internal const string Changelog = "https://github.com/MuertoGB/MacEfiToolkit/blob/main/CHANGELOG.md";
-        internal const string PaypalDonate = "https://www.paypal.com/donate/?hosted_button_id=Z88F3UEZB47SQ";
-        internal const string Email = "mailto:muertogb@proton.me";
-        internal const string FlexBv5 = "https://pldaniels.com/flexbv5/";
-        internal const string GithubHomepage = "https://github.com/MuertoGB/MacEfiToolkit";
-        internal const string GithubManual = "https://github.com/MuertoGB/MacEfiToolkit/blob/main/MANUAL.md#application-manual";
-        internal const string GithubIssues = "https://github.com/MuertoGB/MacEfiToolkit/issues";
-        internal const string GithubLatestVersion = "https://github.com/MuertoGB/MacEfiToolkit/releases/latest";
-        internal const string VersionManifest = "https://raw.githubusercontent.com/MuertoGB/MacEfiToolkit/main/stream/manifests/version.xml";
-
-    }
-
-    internal readonly struct ApplicationVersions
-    {
-        internal const string LZMA_SDK_VERSION = "24.08";
-        internal const string CURRENT_BUILD = "250302.0005";
-        internal const string CURRENT_CHANNEL = "Stable";
-    }
-    #endregion
-
-    #region Enum
+    #region Enums
     internal enum ExitAction
     {
         Restart,
@@ -75,7 +36,6 @@ namespace Mac_EFI_Toolkit
         Error
     }
     #endregion
-
     static class Program
     {
         #region Internal Members
@@ -113,7 +73,7 @@ namespace Mac_EFI_Toolkit
             Application.ApplicationExit += OnExiting;
 
             // Check if the OS is supported (Windows 7 or later is required).
-            if (!IsSupportedOS())
+            if (!SystemTools.IsSupportedOS())
             {
                 return;
             }
@@ -126,7 +86,7 @@ namespace Mac_EFI_Toolkit
             Application.SetCompatibleTextRenderingDefault(false);
 
             // Load custom fonts into memory.
-            if (!LoadCustomFont(Properties.Resources.segmdl2, out Font[] fonts))
+            if (!FontResolver.LoadCustomFont(Properties.Resources.segmdl2, out Font[] fonts))
             {
                 return;
             }
@@ -332,51 +292,6 @@ namespace Mac_EFI_Toolkit
             }
         }
 
-        private static bool IsSupportedOS()
-        {
-            if (SystemTools.GetKernelVersion.ProductMajorPart >= 10)
-            {
-                return true;
-            }
-
-            MessageBox.Show(
-                DIALOGSTRINGS.REQUIRES_WIN_10,
-                DIALOGSTRINGS.UNSUPPORTED_OS,
-                MessageBoxButtons.OK,
-                MessageBoxIcon.Error);
-
-            return false;
-        }
-
-        private static bool LoadCustomFont(byte[] fontbuffer, out Font[] fonts)
-        {
-            fonts = null;
-
-            if (fontbuffer == null)
-            {
-                return false;
-            }
-
-            try
-            {
-                FontFamily ffLoadFont = FontResolver.LoadFont(fontbuffer);
-
-                fonts = new[]
-                {
-                    new Font(ffLoadFont, 10.0F, FontStyle.Regular),
-                    new Font(ffLoadFont, 12.0F, FontStyle.Regular),
-                    new Font(ffLoadFont, 20.0F, FontStyle.Regular)
-                };
-
-                return true;
-            }
-            catch (Exception e)
-            {
-                Logger.WriteErrorLine(nameof(LoadCustomFont), e.GetType(), e.Message);
-                return false;
-            }
-        }
-
         public static void HandleDragEnter(object sender, DragEventArgs e)
         {
             // Check if the dragged data is a file.
@@ -402,21 +317,6 @@ namespace Mac_EFI_Toolkit
 
             // Disable the drop operation.
             e.Effect = DragDropEffects.None;
-        }
-
-        internal static bool IsRunningUnderWine()
-        {
-            string strNtdll = "ntdll.dll";
-            string strWineGetVersion = "wine_get_version";
-
-            IntPtr ptrHandle = NativeMethods.LoadLibrary(strNtdll);
-            if (ptrHandle == IntPtr.Zero)
-            {
-                return false;
-            }
-
-            IntPtr ptrWineHandle = NativeMethods.GetProcAddress(ptrHandle, strWineGetVersion);
-            return ptrWineHandle != IntPtr.Zero; // If this function exists, we're running under Wine.
         }
 
         internal static async Task<VersionResult> CheckForNewVersion()
